Eliminate colon from headings in field lists
Fixes #156.

This also refactors the various HTMLTranslator overrides.
